Bug fixes, schematic deletion
Fixed TileEntities and Schematics clashing with certain mods. Schematics
can now be deleted. They also rotate better. The current (definately not
finished) implementation for building is a bit more efficient.. Sort of.
Fixed a bug where, if a clone was digging, and became hungry, they'd
switch between digging and eating really fast, and just got stuck in a
loop.
make a new built 3.0.8. Hope it works.