Code Cleanup, @Override annotations all checked
Cleaned up some of the code, removed all unused bits. Ensured that
(hopefully) all methods which override something, use the @Override
annotation, to help transitioning to 1.8 a bit easier.
Fixed a bug where when 'Farm' is turned on and a skull is placed by the
clone, the game crashes. Should be fixed..