Fixed some bugs
Should hopefully now work with Ars Magica.
Fixed some other bugs, should crash less. Some GUI's may be missing
certain elements. Lists with a display of the block next to it, may be
missing the block image. Better to be missing an icon, or a name, than
have the whole thing crash IMO